CQ-105332 Kiosk mode reduces touchable area of content (#1)
Problem: The Plugin adds an OverlayView on top of the StatusBar in order to prevent user interactions. Although that's generally a good approach, the OverlayView also prevents the user from interacting with content because the OverlayView is enabled all the time and not just when the StatusBar is visible.

Fix: The Cordova app is most likely in immersive fullscreen mode. We can make use of that and install an `OnSystemUiVisibilityChangeListener`. As soon as the SystemUI is triggered we add the OverlayView that prevents user interactions. After ~2 seconds the System allows us to safely remove the SystemUI again and with that the OverlayView.